requirements.txt
Django==5.1.1
djangorestframework==3.15.2
Pillow==10.4.0
